requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-ipware==2.1.0
idna==2.9
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
